develop/modules/spacetrack-tle-producer/src/get_tle_from_api/requirements.txt
psycopg2-binary
confluent_kafka
PyYAML